+++ app/src/main/java/com/designdemo/uaha/data/model/DeviceEntity.kt
@@ -1,4 +1,4 @@
-package com.designdemo.uaha.data
+package com.designdemo.uaha.data.model
 
 import com.google.gson.annotations.SerializedName
