+++ app/src/main/java/com/designdemo/uaha/data/model/wiki/Main.kt
@@ -1,9 +1,9 @@
-package com.designdemo.uaha.data.wiki
+package com.designdemo.uaha.data.model.wiki
 
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
 
-class Revisions {
+class Main {
 
     @SerializedName("warnings")
     @Expose
