+++ app/src/main/java/com/designdemo/uaha/data/model/wiki/Revisions.kt
@@ -1,9 +1,9 @@
-package com.designdemo.uaha.data.wiki
+package com.designdemo.uaha.data.model.wiki
 
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
 
-class Main {
+class Revisions {
 
     @SerializedName("warnings")
     @Expose
