+++ app/src/main/java/com/designdemo/uaha/net/WikiApiService.kt
@@ -4,7 +4,7 @@ import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
 
-import com.designdemo.uaha.data.wiki.WikiResponse
+import com.designdemo.uaha.data.model.wiki.WikiResponse
 
 interface WikiApiService {
