+++ app/src/main/java/com/designdemo/uaha/view/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.designdemo.uaha.ui
+package com.designdemo.uaha.view
 
 import android.app.SearchManager
 import android.content.Context
