+++ app/src/main/res/layout/activity_bottom_nav.xml
@@ -5,7 +5,7 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.designdemo.uaha.ui.BottomNavActivity">
+    tools:context="com.designdemo.uaha.view.BottomNavActivity">
 
     <TextView
         android:id="@+id/message"
